agents/planning-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "planning-agent"
version = "0.1.0"
description = "Planning and task breakdown agent for multi-agent platform"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore[strands-agents]>=0.1.0",
    "bedrock-agentcore-starter-toolkit>=0.1.0",
    "strands-agents>=0.1.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.39.15",
    "poethepoet>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "black>=23.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
planning-agent = "src.agent:main"

[tool.poe.tasks]
# Development tasks
dev = "python -m src.runtime"
invoke = "agentcore invoke"
launch = "agentcore launch"
test = "pytest tests/"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true